Moving RESIZE constant
