+++ ui/components/ui/textarea/textarea.constants.js
@@ -1,0 +1,8 @@
+export const RESIZE = {
+  NONE: 'none',
+  BOTH: 'both',
+  HORIZONTAL: 'horizontal',
+  VERTICAL: 'vertical',
+  INITIAL: 'initial',
+  INHERIT: 'inherit',
+};
