+++ ui/components/ui/textarea/textarea.js
@@ -4,13 +4,13 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
 import {
-  RESIZE,
   BorderStyle,
-  BLOCK_SIZES,
+  BlockSize,
   Size,
   BorderColor,
   BackgroundColor,
 } from '../../../helpers/constants/design-system';
+import { RESIZE } from './textarea.constants';
 
 import Box from '../box';
 
@@ -40,7 +40,7 @@ const TextArea = ({
       borderRadius={Size.SM}
       borderStyle={BorderStyle.solid}
       padding={4}
-      width={BLOCK_SIZES.FULL}
+      width={BlockSize.Full}
       {...boxProps}
     >
       {(boxClassName) => (
