+++ ui/components/ui/textarea/textarea.stories.js
@@ -2,14 +2,13 @@ import React from 'react';
 import { useArgs } from '@storybook/client-api';
 
 import {
-  RESIZE,
   BorderStyle,
-  BLOCK_SIZES,
+  BlockSize,
   BorderRadius,
   BorderColor,
   Size,
 } from '../../../helpers/constants/design-system';
-
+import { RESIZE } from './textarea.constants';
 import README from './README.mdx';
 import Textarea from '.';
 
@@ -115,6 +114,6 @@ Scrollable.args = {
     borderRadius: BorderRadius.none,
     borderStyle: BorderStyle.none,
     padding: [2, 4],
-    width: BLOCK_SIZES.FULL,
+    width: BlockSize.Full,
   },
 };
